+++ AddOns/MDTGuide/CHANGES.md
@@ -1,3 +1,1 @@
-- Fix zooming to current pull when using enemy forces mode or when auto-hiding is enabled
-- Adjust preset selection dropdown width to make the settings button visible
-- Hide dungeon selection buttons in guide mode
+- Updated ToC version for patch 11.1