+++ AddOns/MDTGuide/CHANGES.md
@@ -1,1 +1,2 @@
-- Updated ToC version for patch 11.1.5
+- Updated ToC version for patch 11.1.7
+- Hide season selection dropdown in guide mode