+++ AddOns/MDTGuide/CHANGES.md
@@ -1,2 +1,1 @@
-- Updated ToC version for patch 11.1.7
-- Hide season selection dropdown in guide mode
+- Updated ToC version for patch 11.2