+++ AddOns/MDTGuide/CHANGES.md
@@ -1,1 +1,2 @@
-- Updated ToC version for patch 11.2
+- Fix boss kill detection in enemy forces mode
+- Color dead enemies of current pull in enemy forces mode