+++ AddOns/MDTGuide/CHANGES.md
@@ -1,1 +1,1 @@
-- Updated ToC version for patch 11.2.5
+- Remove debug code