+++ AddOns/MDTGuide/CHANGES.md
@@ -1,1 +1,1 @@
-- Add workaround for bad zone to dungeon mapping for Academy and Seat
+- Update ToC version for patch 12.0.5